Reorganize build system to use lisp description
The build system is basically just a collection of filters. It seems like lisp
is ideal for working with this kind of problem. I'm only going to implement the
few parts of lisp that I actually need, but hopefully this will end up being a
much more expressive way to generating my site.